sync_db_scripts/requirements.txt
# Зависимости для синхронизации БД

psycopg2-binary==2.9.9  # Драйвер PostgreSQL для Python
